separate scope analysis type contracts
